// src/types/beer-json.ts
export type FermentableBase = {
  name: string;
  type:
    | "dry extract"
    | "extract"
    | "grain"
    | "sugar"
    | "fruit"
    | "juice"
    | "honey"
    | "other";
  origin?: string;
  producer?: string;
  product_id?: string;
  grain_group?:
    | "base"
    | "caramel"
    | "flaked"
    | "roasted"
    | "specialty"
    | "smoked"
    | "adjunct";
  yield: YieldType;
  color: ColorType;
};
// ...
export type FermentableAdditionType = FermentableBase & {
  timing?: TimingType;
  amount: VolumeType | MassType;
};

export type YieldType = {
  fine_grind?: PercentType;
  coarse_grind?: PercentType;
  fine_coarse_difference?: PercentType;
  potential?: GravityType;
};
// ...
export type VolumeType = {
  unit: VolumeUnitType;
  value: number;
};

export type MassType = {
  unit: MassUnitType;
  value: number;
};

export type DiastaticPowerType = {
  unit: DiastaticPowerUnitType;
  value: number;
};

export type TemperatureType = {
  unit: TemperatureUnitType;
  value: number;
};

export type PressureType = {
  unit: PressureUnitType;
  value: number;
};

export type AcidityType = {
  unit: AcidityUnitType;
  value: number;
};

export type TimeType = {
  unit: TimeUnitType;
  value: number;
};

export type ColorType = {
  unit: ColorUnitType;
  value: number;
};

export type CarbonationType = {
  unit: CarbonationUnitType;
  value: number;
};

export type BitternessType = {
  unit: BitternessUnitType;
  value: number;
};

export type GravityType = {
  unit: GravityUnitType;
  value: number;
};
// ...
export type VolumeUnitType =
  | "ml"
  | "l"
  | "tsp"
  | "tbsp"
  | "floz"
  | "cup"
  | "pt"
  | "qt"
  | "gal"
  | "bbl"
  | "ifloz"
  | "ipt"
  | "iqt"
  | "igal"
  | "ibbl";
export type MassUnitType = "mg" | "g" | "kg" | "lb" | "oz";
// ...
export type UseType =
  | "add_to_mash"
  | "add_to_boil"
  | "add_to_fermentation"
  | "add_to_package";
export type TimingType = {
  time?: TimeType;
  duration?: TimeType;
  continuous?: boolean;
  specific_gravity?: GravityType;
  pH?: AcidityType;
  step?: number;
  use?: UseType;
};